# contracts/openapi/specs/v1/auth.yaml
openapi: 3.1.0
info:
  title: QTicket Auth API
  description: Authentication
  version: 1.0.0
servers:
  - url: /api/v1/auth

paths:
  /.well-known/jwks.json:
    get:
      summary: Get JWKS for JWT verification
      operationId: getJwks
      tags: [ Auth ]
      responses:
        '200':
          description: Successful JWKS retrieval
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      $ref: '#/components/schemas/Jwk'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /login:
    post:
      summary: Login with email/username and password
      operationId: login
      tags: [ Auth ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: 2FA required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwoFaChallenge'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  /login/2fa:
    post:
      summary: Complete login with second factor
      operationId: verify2fa
      tags: [ Auth ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFaVerifyRequest'
      responses:
        '200':
          description: Successful login after 2FA
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid or expired 2FA challenge
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  /token/refresh:
    post:
      summary: Refresh access token
      operationId: refreshToken
      tags: [ Auth ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Successful token refresh
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Not authenticated
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Refresh token expired
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  /logout:
    post:
      summary: Logout with refresh token
      operationId: logout
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: Refresh token
      responses:
        '204':
          description: Successfully logged out
        '401':
          description: Not authenticated
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProblemDetail:
      $ref: './error.yaml#/components/schemas/ProblemDetail'

    Jwk:
      type: object
      required: [ kty, kid, n, e ]
      properties:
        kty:
          type: string
          description: Key type
        kid:
          type: string
          description: Key ID
        use:
          type: string
          description: Key use
        alg:
          type: string
          description: Algorithm
        n:
          type: string
          description: Modulus
        e:
          type: string
          description: Exponent
    # Login schema
    LoginRequest:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
          description: Email or username
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [ accessToken, refreshToken, expiresIn ]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          format: int64
          description: Access token lifetime in seconds

    # Two-factor challenge schema
    TwoFaMethod:
      type: string
      enum: [ totp, passkey ]

    TwoFaChallenge:
      type: object
      required: [ challengeId, methods ]
      properties:
        methods:
          type: array
          description: Supported methods for two-factor authentication
          items:
            $ref: '#/components/schemas/TwoFaMethod'
        challengeId:
          type: string
          description: Challenge ID for two-factor authentication
        # For passkey I'll add a WebAuthn challenge object here
        publicKeyCredentialRequestOptions:
          type: object
          description: WebAuthn options for passkey challenges (opaque to clients you control)

    # Two-factor verification schema
    TwoFaVerifyRequest:
      type: object
      required: [ challengeId, method ]
      properties:
        challengeId:
          type: string
          description: Challenge ID for two-factor authentication
        method:
          $ref: '#/components/schemas/TwoFaMethod'
        code:
          type: string
          description: Code for two-factor authentication, for example a TOTP code
        passkeyAssertion:
          type: string
          description: Passkey assertion for two-factor authentication

    # Refresh token schema
    RefreshTokenRequest:
      type: object
      required: [ refreshToken ]
      properties:
        refreshToken:
          type: string
          description: Refresh token
          